chore: Restructure
