+++ frontend/app/(home)/page.tsx
@@ -1,0 +1,10 @@
+import Link from "next/link";
+import Hero from "./Hero";
+
+export default function HomePage() {
+  return (
+    <main className="">
+      <Hero />
+    </main>
+  );
+}
